Split shop mutations into separate files
