+++ saleor/graphql/shop/mutations/__init__.py
@@ -1,0 +1,29 @@
+from .attributes import (
+    CategoryAttributeSettingsUpdate,
+    CategorySettingsReorderAttributes,
+)
+from .notifications import (
+    StaffNotificationRecipientCreate,
+    StaffNotificationRecipientDelete,
+    StaffNotificationRecipientUpdate,
+)
+from .settings import (
+    OrderSettingsUpdate,
+    ShopAddressUpdate,
+    ShopDomainUpdate,
+    ShopFetchTaxRates,
+    ShopSettingsUpdate,
+)
+
+__all__ = [
+    "CategoryAttributeSettingsUpdate",
+    "CategorySettingsReorderAttributes",
+    "OrderSettingsUpdate",
+    "ShopAddressUpdate",
+    "ShopDomainUpdate",
+    "ShopFetchTaxRates",
+    "ShopSettingsUpdate",
+    "StaffNotificationRecipientCreate",
+    "StaffNotificationRecipientDelete",
+    "StaffNotificationRecipientUpdate",
+]
