+++ saleor/graphql/shop/mutations/settings.py
@@ -1,0 +1,209 @@
+import graphene
+from django.core.exceptions import ValidationError
+
+from ....account import models as account_models
+from ....core.error_codes import ShopErrorCode
+from ....core.permissions import OrderPermissions, SitePermissions
+from ....core.utils.url import validate_storefront_url
+from ...account.i18n import I18nMixin
+from ...account.types import AddressInput
+from ...core.enums import WeightUnitsEnum
+from ...core.mutations import BaseMutation
+from ...core.types.common import OrderSettingsError, ShopError
+from ..types import OrderSettings, Shop
+
+
+class ShopSettingsInput(graphene.InputObjectType):
+    header_text = graphene.String(description="Header text.")
+    description = graphene.String(description="SEO description.")
+    include_taxes_in_prices = graphene.Boolean(description="Include taxes in prices.")
+    display_gross_prices = graphene.Boolean(
+        description="Display prices with tax in store."
+    )
+    charge_taxes_on_shipping = graphene.Boolean(description="Charge taxes on shipping.")
+    track_inventory_by_default = graphene.Boolean(
+        description="Enable inventory tracking."
+    )
+    default_weight_unit = WeightUnitsEnum(description="Default weight unit.")
+    automatic_fulfillment_digital_products = graphene.Boolean(
+        description="Enable automatic fulfillment for all digital products."
+    )
+    default_digital_max_downloads = graphene.Int(
+        description="Default number of max downloads per digital content URL."
+    )
+    default_digital_url_valid_days = graphene.Int(
+        description="Default number of days which digital content URL will be valid."
+    )
+    default_mail_sender_name = graphene.String(
+        description="Default email sender's name."
+    )
+    default_mail_sender_address = graphene.String(
+        description="Default email sender's address."
+    )
+    customer_set_password_url = graphene.String(
+        description="URL of a view where customers can set their password."
+    )
+
+
+class SiteDomainInput(graphene.InputObjectType):
+    domain = graphene.String(description="Domain name for shop.")
+    name = graphene.String(description="Shop site name.")
+
+
+class ShopSettingsUpdate(BaseMutation):
+    shop = graphene.Field(Shop, description="Updated shop.")
+
+    class Arguments:
+        input = ShopSettingsInput(
+            description="Fields required to update shop settings.", required=True
+        )
+
+    class Meta:
+        description = "Updates shop settings."
+        permissions = (SitePermissions.MANAGE_SETTINGS,)
+        error_type_class = ShopError
+        error_type_field = "shop_errors"
+
+    @classmethod
+    def clean_input(cls, _info, _instance, data):
+        if data.get("customer_set_password_url"):
+            try:
+                validate_storefront_url(data["customer_set_password_url"])
+            except ValidationError as error:
+                raise ValidationError(
+                    {"customer_set_password_url": error}, code=ShopErrorCode.INVALID
+                )
+        return data
+
+    @classmethod
+    def construct_instance(cls, instance, cleaned_data):
+        for field_name, desired_value in cleaned_data.items():
+            current_value = getattr(instance, field_name)
+            if current_value != desired_value:
+                setattr(instance, field_name, desired_value)
+        return instance
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        instance = info.context.site.settings
+        data = data.get("input")
+        cleaned_input = cls.clean_input(info, instance, data)
+        instance = cls.construct_instance(instance, cleaned_input)
+        cls.clean_instance(info, instance)
+        instance.save()
+        return ShopSettingsUpdate(shop=Shop())
+
+
+class ShopAddressUpdate(BaseMutation, I18nMixin):
+    shop = graphene.Field(Shop, description="Updated shop.")
+
+    class Arguments:
+        input = AddressInput(description="Fields required to update shop address.")
+
+    class Meta:
+        description = (
+            "Update the shop's address. If the `null` value is passed, the currently "
+            "selected address will be deleted."
+        )
+        permissions = (SitePermissions.MANAGE_SETTINGS,)
+        error_type_class = ShopError
+        error_type_field = "shop_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        site_settings = info.context.site.settings
+        data = data.get("input")
+
+        if data:
+            if not site_settings.company_address:
+                company_address = account_models.Address()
+            else:
+                company_address = site_settings.company_address
+            company_address = cls.validate_address(data, company_address, info=info)
+            company_address.save()
+            site_settings.company_address = company_address
+            site_settings.save(update_fields=["company_address"])
+        else:
+            if site_settings.company_address:
+                site_settings.company_address.delete()
+        return ShopAddressUpdate(shop=Shop())
+
+
+class ShopDomainUpdate(BaseMutation):
+    shop = graphene.Field(Shop, description="Updated shop.")
+
+    class Arguments:
+        input = SiteDomainInput(description="Fields required to update site.")
+
+    class Meta:
+        description = "Updates site domain of the shop."
+        permissions = (SitePermissions.MANAGE_SETTINGS,)
+        error_type_class = ShopError
+        error_type_field = "shop_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        site = info.context.site
+        data = data.get("input")
+        domain = data.get("domain")
+        name = data.get("name")
+        if domain is not None:
+            site.domain = domain
+        if name is not None:
+            site.name = name
+        cls.clean_instance(info, site)
+        site.save()
+        return ShopDomainUpdate(shop=Shop())
+
+
+class ShopFetchTaxRates(BaseMutation):
+    shop = graphene.Field(Shop, description="Updated shop.")
+
+    class Meta:
+        description = "Fetch tax rates."
+        permissions = (SitePermissions.MANAGE_SETTINGS,)
+        error_type_class = ShopError
+        error_type_field = "shop_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info):
+        if not info.context.plugins.fetch_taxes_data():
+            raise ValidationError(
+                "Could not fetch tax rates. Make sure you have supplied a "
+                "valid credential for your tax plugin.",
+                code=ShopErrorCode.CANNOT_FETCH_TAX_RATES.value,
+            )
+        return ShopFetchTaxRates(shop=Shop())
+
+
+class OrderSettingsUpdateInput(graphene.InputObjectType):
+    automatically_confirm_all_new_orders = graphene.Boolean(
+        required=True,
+        description="When disabled, all new orders from checkout "
+        "will be marked as unconfirmed. When enabled orders from checkout will "
+        "become unfulfilled immediately.",
+    )
+
+
+class OrderSettingsUpdate(BaseMutation):
+    order_settings = graphene.Field(OrderSettings, description="Order settings.")
+
+    class Arguments:
+        input = OrderSettingsUpdateInput(
+            required=True, description="Fields required to update shop order settings."
+        )
+
+    class Meta:
+        description = "Update shop order settings."
+        permissions = (OrderPermissions.MANAGE_ORDERS,)
+        error_type_class = OrderSettingsError
+        error_type_field = "order_settings_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        instance = info.context.site.settings
+        instance.automatically_confirm_all_new_orders = data["input"][
+            "automatically_confirm_all_new_orders"
+        ]
+        instance.save(update_fields=["automatically_confirm_all_new_orders"])
+        return OrderSettingsUpdate(order_settings=instance)
