+++ saleor/graphql/shop/tests/benchmark/test_category_settings.py
@@ -18,7 +18,7 @@ def test_category_settings_update_by_staff(
 ):
     query = """
         mutation CategoryAttributeSettingsUpdate(
-            $input: CategoryAttributeSettingsInput!
+            $input: AttributeSettingsInput!
         ) {
             categoryAttributeSettingsUpdate(input: $input) {
                 categoryAttributeSettings {
